requirements.txt
colorama==0.4.6
discord.py==2.3.2
Pillow==10.0.0
Requests==2.31.0
